backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.39.0
pandas==2.2.3
jinja2==3.1.4
mail-parser==3.15.0
python-multipart==0.0.12
